When retrying messages, ensure that the original message is retried.
Previously, modifications to the message made by filters or processors
were retained when the message was requeued. This caused unexpected
results when the message was modified. As a result, message filters
and DLQ processing did not work together. This commit fixes the issue
by passing a deep copy of the message to the filter chain, so that the
original message can be retried.